requirements.txt
ultralytics==8.3.0
torch
opencv-python
PyYAML
requests
fastapi
uvicorn
python-dotenv
matplotlib
pandas